moved LocalRegistryStore to local
