+++ sdk/python/feast/infra/local.py
@@ -19,7 +19,7 @@ from feast.infra.provider import (
 )
 from feast.protos.feast.types.EntityKey_pb2 import EntityKey as EntityKeyProto
 from feast.protos.feast.types.Value_pb2 import Value as ValueProto
-from feast.registry import Registry
+from feast.registry import Registry, RegistryStore
 from feast.repo_config import RepoConfig
 
 
@@ -152,3 +152,39 @@ def _to_naive_utc(ts: datetime):
         return ts
     else:
         return ts.astimezone(pytz.utc).replace(tzinfo=None)
+
+class LocalRegistryStore(RegistryStore):
+    def __init__(self, repo_path: Path, registry_path_string: str):
+        registry_path = Path(registry_path_string)
+        if registry_path.is_absolute():
+            self._filepath = registry_path
+        else:
+            self._filepath = repo_path.joinpath(registry_path)
+
+    def get_registry_proto(self):
+        registry_proto = RegistryProto()
+        if self._filepath.exists():
+            registry_proto.ParseFromString(self._filepath.read_bytes())
+            return registry_proto
+        raise FileNotFoundError(
+            f'Registry not found at path "{self._filepath}". Have you run "feast apply"?'
+        )
+
+    def update_registry_proto(self, registry_proto: RegistryProto):
+        self._write_registry(registry_proto)
+
+    def teardown(self):
+        try:
+            self._filepath.unlink()
+        except FileNotFoundError:
+            # If the file deletion fails with FileNotFoundError, the file has already
+            # been deleted.
+            pass
+
+    def _write_registry(self, registry_proto: RegistryProto):
+        registry_proto.version_id = str(uuid.uuid4())
+        registry_proto.last_updated.FromDatetime(datetime.utcnow())
+        file_dir = self._filepath.parent
+        file_dir.mkdir(exist_ok=True)
+        self._filepath.write_bytes(registry_proto.SerializeToString())
+
